chore(rpc): remove many unused code about tx generation
