+++ commands/box/wallet/root.go
@@ -117,26 +117,7 @@ func init() {
 			Short: "List transactions for an account",
 			Run:   listTransactionsCmdFunc,
 		},
-		&cobra.Command{
-			Use:   "console",
-			Short: "Start wallet interactive console",
-			Run:   walletConsoleCmdFunc,
-		},
 	)
-}
-
-func walletConsoleCmdFunc(cmd *cobra.Command, args []string) {
-	wltMgr, err := wallet.NewWalletManager(walletDir)
-	if err != nil {
-		fmt.Println(err.Error())
-		return
-	}
-	for _, acc := range wltMgr.ListAccounts() {
-		fmt.Println("Managed Address:", acc.Addr(), "Public Key Hash:", hex.EncodeToString(acc.PubKeyHash()))
-	}
-	cli := client.NewConnectionWithViper(viper.GetViper())
-	defer cli.Close()
-	startConsole(cli, wltMgr)
 }
 
 func newAccountCmdFunc(cmd *cobra.Command, args []string) {
