+++ core/txlogic/tx_utils_test.go
@@ -1,15 +1,20 @@
+// Copyright (c) 2018 ContentBox Authors.
+// Use of this source code is governed by a MIT-style
+// license that can be found in the LICENSE file.
+
 package txlogic
 
 import (
 	"testing"
 
+	"github.com/BOXFoundation/boxd/core/types"
 	"github.com/BOXFoundation/boxd/script"
 )
 
 func TestNewIssueTokenUtxoWrap(t *testing.T) {
 	addr := "b1ndoQmEd83y4Fza5PzbUQDYpT3mV772J5o"
 	name, sym, deci := "box token", "BOX", uint8(8)
-	tag := NewTokenTag(name, sym, deci)
+	tag := types.NewTokenTag(name, sym, deci)
 	uw, err := NewIssueTokenUtxoWrap(addr, tag, 1, 10000)
 	if err != nil {
 		t.Fatal(err)
