+++ integration_tests/test_token.go
@@ -15,7 +15,6 @@ import (
 	"github.com/BOXFoundation/boxd/core/types"
 	"github.com/BOXFoundation/boxd/integration_tests/utils"
 	"github.com/BOXFoundation/boxd/rpc/client"
-	"google.golang.org/grpc"
 )
 
 // TokenTest manage circulation of token
@@ -74,14 +73,18 @@ func (t *TokenTest) HandleFunc(addrs []string, index *int) (exit bool) {
 		return true
 	}
 	issuer, sender, receivers := addrs[0], addrs[1], addrs[2:]
-	tx, _, _, err := utils.NewTx(AddrToAcc[miner], []string{issuer, sender},
-		[]uint64{subsidy, testFee}, peerAddr)
+	conn, err := client.GetGRPCConn(peerAddr)
+	if err != nil {
+		logger.Warn(err)
+		return false
+	}
+	defer conn.Close()
+	tx, _, _, err := client.NewTx(AddrToAcc[miner], []string{issuer, sender},
+		[]uint64{subsidy, testFee}, conn)
 	if err != nil {
 		logger.Error(err)
 		return
 	}
-	conn, _ := grpc.Dial(peerAddr, grpc.WithInsecure())
-	defer conn.Close()
 	if err := client.SendTransaction(conn, tx); err != nil &&
 		!strings.Contains(err.Error(), core.ErrOrphanTransaction.Error()) {
 		logger.Error(err)
@@ -101,15 +104,18 @@ func tokenRepeatTest(issuer, sender string, receivers []string,
 	logger.Info("=== RUN   tokenRepeatTest")
 	defer logger.Info("=== DONE   tokenRepeatTest")
 
-	conn, _ := grpc.Dial(peerAddr, grpc.WithInsecure())
+	conn, err := client.GetGRPCConn(peerAddr)
+	if err != nil {
+		logger.Panic(err)
+	}
 	defer conn.Close()
 
 	// issue some token
 	totalSupply := uint64(10000)
 	logger.Infof("%s issue %d token to %s", issuer, totalSupply, sender)
 
-	tx, tid, _, err := client.NewIssueTokenTx(conn, AddrToAcc[issuer], sender,
-		tag, totalSupply)
+	tx, tid, _, err := client.NewIssueTokenTx(AddrToAcc[issuer], sender, tag,
+		totalSupply, conn)
 	if err != nil {
 		logger.Panic(err)
 	}
@@ -139,8 +145,8 @@ func tokenRepeatTest(issuer, sender string, receivers []string,
 	logger.Infof("start to create %d token txs from %s to %s on %s",
 		times, sender, receiver, peerAddr)
 	txTotalAmount := totalAmount/2 + uint64(rand.Int63n(int64(totalAmount)/2))
-	txs, err := utils.NewTokenTxs(AddrToAcc[sender], receiver, txTotalAmount, times,
-		tid, peerAddr)
+	txs, err := client.NewTokenTxs(AddrToAcc[sender], receiver, txTotalAmount, times,
+		tid, conn)
 	if err != nil {
 		logger.Panic(err)
 	}
